Break apart the Gossip.Games module
